Seperate admin panel from application.rb removed inheritence
